requirements.txt
# Core dependencies for mtrack experiment tracking

# Configuration
pyyaml>=6.0

# Hugging Face ecosystem
transformers>=4.30.0
huggingface-hub>=0.16.0
datasets>=2.14.0

# ML frameworks (choose based on your needs)
torch>=2.0.0
torchvision>=0.15.0

# Utilities
numpy>=1.24.0
